move images/ to assets/
