+++ src/assets/images/brand-text.png
