+++ src/assets/images/index.js
