+++ src/assets/images/kaltura-logo.png
