+++ src/assets/images/logo-outline.png
