+++ src/components/Footer/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { Link } from 'react-router-dom'
 import { Grid, Image } from 'semantic-ui-react'
 import './index.css'
-import { logoOutline as logo } from '../../images'
+import { logoOutline as logo } from '../../assets/images'
 import { util } from '../../utils'
 
 export function ClassTranscribeFooter() {
